fix: move PermissionEditor to permissions module
